Extract filters to a Svelte component
