+++ src/lib/browse/data.ts
@@ -9,25 +9,26 @@ export interface Scheme {
 }
 
 // Takes a GeoJSON file representing a bunch of scheme files combined into one.
-// Cleans up some problems with this input (temporary, until fixed upstream)
-// and returns a list of Schemes. Each feature (intervention) in the GJ links
-// back to one of these schemes by scheme_reference.
-export function processInput(gj: FeatureCollection): Scheme[] {
-  let byScheme = {};
+// Cleans up some problems with this input (temporary, until fixed upstream) and
+// returns a dictionary of Schemes, keyed (and ordered) by scheme_reference.
+// Each feature (intervention) in the GJ links back to one of these schemes by
+// scheme_reference.
+export function processInput(gj: FeatureCollection): Map<string, Scheme> {
+  let schemes = new Map();
 
   // Assume the input has a top-level dictionary keyed by scheme_reference
   for (let [scheme_reference, scheme] of Object.entries(gj.schemes)) {
-    byScheme[scheme_reference] = {
+    schemes.set(scheme_reference, {
       scheme_reference,
       num_features: 0,
       authority_or_region: scheme.authority_or_region,
       capital_scheme_id: scheme.capital_scheme_id,
       funding_programme: scheme.funding_programme,
-    };
+    });
   }
 
   for (let feature of gj.features) {
-    byScheme[feature.properties.scheme_reference].num_features++;
+    schemes.get(feature.properties.scheme_reference).num_features++;
     // Remove extraneous fields from the input. Aside from the v1
     // InterventionProps, there should just be a scheme_reference linking to
     // the top-level dictionary.
@@ -40,5 +41,6 @@ export function processInput(gj: FeatureCollection): Scheme[] {
     delete feature.properties.centroid_lat;
   }
 
-  return Object.values(byScheme);
+  window.x = schemes;
+  return schemes;
 }
