+++ src/pages/BrowseSchemes.svelte
@@ -1,26 +1,20 @@
 <script lang="ts">
-  // @ts-nocheck
-  // TODO After figuring out the features of this page, work on the errors
-  // here. Need to decide how much we use gjScheme, and what new feature-level
-  // properties to expect.
   import { initAll } from "govuk-frontend";
   import "../style/main.css";
   import type { GeoJSON } from "geojson";
   import { onDestroy, onMount } from "svelte";
   import BaselayerSwitcher from "../lib/BaselayerSwitcher.svelte";
   import { processInput, type Scheme } from "../lib/browse/data";
+  import Filters from "../lib/browse/Filters.svelte";
   import SchemeCard from "../lib/browse/SchemeCard.svelte";
-  import CollapsibleCard from "../lib/common/CollapsibleCard.svelte";
   import FileInput from "../lib/common/FileInput.svelte";
   import Layout from "../lib/common/Layout.svelte";
   import MapTooltips from "../lib/common/MapTooltips.svelte";
   import InterventionLayer from "../lib/draw/InterventionLayer.svelte";
   import ErrorMessage from "../lib/govuk/ErrorMessage.svelte";
-  import FormElement from "../lib/govuk/FormElement.svelte";
   import SecondaryButton from "../lib/govuk/SecondaryButton.svelte";
-  import Select from "../lib/govuk/Select.svelte";
   import Legend from "../lib/Legend.svelte";
-  import Map from "../lib/Map.svelte";
+  import MapLibreMap from "../lib/Map.svelte";
   import ZoomOutMap from "../lib/ZoomOutMap.svelte";
   import { bbox, prettyPrintMeters } from "../maplibre_helpers";
   import { gjScheme, map } from "../stores";
@@ -35,86 +29,12 @@
   const schema = "v1";
   let errorMessage = "";
 
-  // unique by scheme_reference
-  let schemes: Scheme[] = [];
-  let filterText = "";
-  // by scheme_reference
+  let schemes: Map<string, Scheme> = new Map();
   let showSchemes: Set<string> = new Set();
-
-  // Stats about filtered schemes
-  // TODO Extract into a component
-  let counts = { area: 0, route: 0, crossing: 0, other: 0 };
-
-  // Dropdown filters
-  let authorities: [string, string][] = [];
-  let filterAuthority = "";
-  let fundingProgrammes: [string, string][] = [];
-  let filterFundingProgramme = "";
 
   onDestroy(() => {
     gjScheme.set(null);
   });
-
-  $: if ($gjScheme) {
-    // When schemes or filter changes, update showSchemes
-    showSchemes.clear();
-    if (filterText || filterAuthority || filterFundingProgramme) {
-      let filterNormalized = filterText.toLowerCase();
-      for (let feature of $gjScheme.features) {
-        // TODO This is a very blunt free-form text search, of any property.
-        if (
-          !JSON.stringify(feature.properties)
-            .toLowerCase()
-            .includes(filterNormalized)
-        ) {
-          continue;
-        }
-        if (
-          filterAuthority &&
-          feature.properties.authority_or_region != filterAuthority
-        ) {
-          continue;
-        }
-        if (
-          filterFundingProgramme &&
-          feature.properties.funding_programme != filterFundingProgramme
-        ) {
-          continue;
-        }
-        showSchemes.add(feature.properties.scheme_reference);
-      }
-    } else {
-      for (let scheme of schemes) {
-        showSchemes.add(scheme.scheme_reference);
-      }
-    }
-
-    // Hide things on the map
-    gjScheme.update((gj) => {
-      if (!gj) {
-        return null;
-      }
-      for (let feature of gj.features) {
-        if (showSchemes.has(feature.properties.scheme_reference)) {
-          delete feature.properties.hide_while_editing;
-        } else {
-          feature.properties.hide_while_editing = true;
-        }
-      }
-      return gj;
-    });
-
-    // Recalculate stats
-    counts = { area: 0, route: 0, crossing: 0, other: 0 };
-    for (let feature of $gjScheme?.features) {
-      if (showSchemes.has(feature.properties.scheme_reference)) {
-        counts[feature.properties.intervention_type]++;
-      }
-    }
-
-    // Make Svelte see the update
-    showSchemes = showSchemes;
-  }
 
   function loadFile(text: string) {
     try {
@@ -122,22 +42,6 @@
       schemes = processInput(gj);
       gjScheme.set(gj);
       errorMessage = "";
-
-      // Set up filters
-      let set1 = new Set();
-      let set2 = new Set();
-      for (let x of schemes) {
-        if (x.authority_or_region) {
-          set1.add(x.authority_or_region);
-        }
-        if (x.funding_programme) {
-          set2.add(x.funding_programme);
-        }
-      }
-      authorities = Array.from(set1.entries());
-      authorities.sort();
-      fundingProgrammes = Array.from(set2.entries());
-      fundingProgrammes.sort();
 
       $map?.fitBounds(bbox(gj), { padding: 20, animate: false });
     } catch (err) {
@@ -174,39 +78,12 @@
     {/if}
     <FileInput label="Load from GeoJSON" id="load-geojson" {loadFile} />
 
-    <CollapsibleCard label="Filters">
-      <Select
-        label="Authority or region"
-        id="filterAuthority"
-        choices={authorities}
-        bind:value={filterAuthority}
-      />
-      <Select
-        label="Funding programme"
-        id="filterFundingProgramme"
-        choices={fundingProgrammes}
-        bind:value={filterFundingProgramme}
-      />
-      <FormElement label="Any field" id="filterText">
-        <input
-          type="text"
-          class="govuk-input govuk-input--width-10"
-          id="filterText"
-          bind:value={filterText}
-        />
-        <SecondaryButton on:click={() => (filterText = "")}>
-          Clear
-        </SecondaryButton>
-      </FormElement>
-    </CollapsibleCard>
-
-    <p>
-      Showing {showSchemes.size} schemes ({counts.route} routes, {counts.area} areas,
-      {counts.crossing} crossings, {counts.other} other)
-    </p>
+    {#if schemes.size > 0}
+      <Filters {schemes} bind:showSchemes />
+    {/if}
 
     <ul>
-      {#each schemes as scheme}
+      {#each schemes.values() as scheme}
         {#if showSchemes.has(scheme.scheme_reference)}
           <SchemeCard {scheme} />
         {/if}
@@ -214,7 +91,7 @@
     </ul>
   </div>
   <div slot="main">
-    <Map {style}>
+    <MapLibreMap {style}>
       <InterventionLayer {schema} />
       <BaselayerSwitcher {style} />
       <Legend {schema} />
@@ -226,6 +103,6 @@
         ]}
         contents={tooltip}
       />
-    </Map>
+    </MapLibreMap>
   </div>
 </Layout>
